admin auth

+++ src/modules/products/products.service.ts
@@ -1,10 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Products, ProductsDocument } from '../../schemas/products.schema';
 import { Category, CategoryDocument } from '../../schemas/category.schema';
 import { CreateProductsDto } from '../../dto/create-products.dto';
-import { GetProductsQueryDto } from 'src/dto/query-products.dto';
+import {
+  GetProductsQueryDto,
+  GetProductsQuery,
+} from 'src/dto/query-products.dto';
+import { GetStatus } from 'utils/types';
 
 @Injectable()
 export class ProductsService {
@@ -18,7 +22,7 @@ export class ProductsService {
   }
 
   async findAllProducts(
-    query: GetProductsQueryDto,
+    query: GetProductsQuery,
   ): Promise<{ categoryInfo: Category; productsList: Products[] }[]> {
     const categoryList = await this.categoryModel.find({
       is_active: true,
@@ -34,8 +38,8 @@ export class ProductsService {
     const productsListbyCategories = [];
     for (let i = 0; i < categoryList?.length; i++) {
       const products = await this.productModel
-        .find({ categoryId: categoryList[i]._id })
-        .limit(query?.limit);
+        .find({ categoryId: categoryList[i]._id, isActive: true })
+        .limit(+query?.limit);
       const newList = {
         categoryInfo: categoryList[i],
         productsList: products,
@@ -63,16 +67,47 @@ export class ProductsService {
     query: GetProductsQueryDto,
   ): Promise<Products[]> {
     return await this.productModel
-      .find({ categoryId: catId })
+      .find({ categoryId: catId, isActive: true })
       .skip(query?.skip)
       .limit(query?.limit);
   }
 
-  async delete(id: string): Promise<Products> {
-    return await this.productModel.findByIdAndRemove(id);
+  async delete(id: string): Promise<Products | GetStatus> {
+    const product = await this.productModel.findById(id);
+    if (product) {
+      await this.productModel.findByIdAndRemove(id);
+      return {
+        status: 'Success',
+        message: 'Delete Product Successful',
+      };
+    }
+    throw new HttpException('Product not found', HttpStatus.BAD_REQUEST);
   }
 
   async update(id: string, item: CreateProductsDto): Promise<Products> {
-    return await this.productModel.findByIdAndUpdate(id, item, { new: true });
+    const product = await this.productModel.findById(id);
+    if (product) {
+      return await this.productModel.findByIdAndUpdate(id, item, {
+        new: true,
+      });
+    }
+    throw new HttpException('Product not found', HttpStatus.BAD_REQUEST);
+  }
+
+  async disableProduct(id: string): Promise<Products | GetStatus> {
+    const isDisable = {
+      isActive: false,
+    };
+    const product = await this.productModel.findById(id);
+    if (product) {
+      await this.productModel.findByIdAndUpdate(id, isDisable, {
+        new: true,
+      });
+      return {
+        status: 'Success',
+        message: 'Deactive Product Successful',
+      };
+    }
+    throw new HttpException('Product not found', HttpStatus.BAD_REQUEST);
   }
 }
